Lazy run interval
Do not run interval on module import, but when we call the function. 

Some SPA frameworks (like Angular) detect this timer call and run a change detection every 250ms. Now we run the timer only when we call the function, we have a way to say to the framework "execute the speedometer function but don't check the DOM when `setInterval` is fired."